+++ pom.xml
@@ -8,7 +8,7 @@
   <description>portal</description>
     <properties>
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-      <revision>2.7-SNAPSHOT</revision>
+      <revision>2.6-SNAPSHOT</revision>
     </properties>
     <build>
         <plugins>
