+++ src/main/java/in/succinct/beckn/portal/controller/BapController.java
@@ -1,27 +1,38 @@
 package in.succinct.beckn.portal.controller;
 
 import com.venky.core.string.StringUtil;
+import com.venky.core.util.ObjectUtil;
 import com.venky.swf.controller.Controller;
 import com.venky.swf.controller.annotations.RequireLogin;
+import com.venky.swf.db.Database;
 import com.venky.swf.db.annotations.column.ui.mimes.MimeType;
+import com.venky.swf.db.model.io.json.JSONFormatter;
+import com.venky.swf.db.model.reflection.ModelReflector;
 import com.venky.swf.path.Path;
+import com.venky.swf.controller.TemplatedController;
 import com.venky.swf.routing.Config;
+import com.venky.swf.sql.Expression;
+import com.venky.swf.sql.Operator;
+import com.venky.swf.sql.Select;
 import com.venky.swf.views.BytesView;
 import com.venky.swf.views.View;
 import in.succinct.beckn.Acknowledgement;
 import in.succinct.beckn.Acknowledgement.Status;
 import in.succinct.beckn.Error;
-import in.succinct.beckn.Options;
 import in.succinct.beckn.Request;
 import in.succinct.beckn.Response;
-import org.json.simple.JSONArray;
+import in.succinct.beckn.portal.db.model.api.ApiCall;
+import org.json.simple.JSONObject;
+import org.json.simple.JSONValue;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.List;
+import java.util.UUID;
 
-public class LocalDeliveryBppController extends Controller {
-    public LocalDeliveryBppController(Path path) {
+public class BapController extends Controller {
+    public BapController(Path path) {
         super(path);
     }
     public View nack(Request request, String realm){
@@ -40,23 +51,53 @@ public class LocalDeliveryBppController extends Controller {
     }
     public View ack(Request request){
         Acknowledgement ack = new Acknowledgement(Status.ACK);
-        //ack.setSignature(Request.generateSignature(request.hash(),request.getPrivateKey(request.getContext().getBppId(),request.getContext().getBppId() +".k1")));
+        //ack.setSignature(Request.generateSignature(request.hash(),request.getPrivateKey(request.getContext().getBapId(),request.getContext().getBapId() +".k1")));
         return new BytesView(getPath(),new Response(request.getContext(),ack).toString().getBytes(StandardCharsets.UTF_8));
     }
 
 
     private View act(){
-        Request request = null;
+        Request request =null ;
         try {
             request = new Request(StringUtil.read(getPath().getInputStream()));
             if (!Config.instance().getBooleanProperty("beckn.auth.enabled", false)  || request.verifySignature("Authorization",getPath().getHeaders())){
+                String messageId = request.getContext().getMessageId();
+                if (ObjectUtil.isVoid(messageId)){
+                    messageId = UUID.randomUUID().toString();
+                }
+                JSONFormatter formatter = new JSONFormatter();
+                List<ApiCall> apiCalls = new Select(true,true).from(ApiCall.class).where(new Expression(ModelReflector.instance(ApiCall.class).getPool(),"MESSAGE_ID", Operator.EQ,messageId)).execute();
+                ApiCall apiCall =null;
+                if (apiCalls.isEmpty()){
+                    apiCall = Database.getTable(ApiCall.class).newRecord();
+                    apiCall.setMessageId(messageId);
+                }else {
+                    apiCall  = apiCalls.get(0);
+                }
+
+                JSONObject responseCollection = new JSONObject();
+                JSONObject headersCollection = new JSONObject();
+
+                if (!apiCall.getReflector().isVoid(apiCall.getCallBackPayload())){
+                    responseCollection = (JSONObject) JSONValue.parse(apiCall.getCallBackPayload());
+                    headersCollection = (JSONObject) JSONValue.parse(apiCall.getCallBackHeaders());
+                }
+
+                responseCollection.put(request.getContext().getBppId(),request.getInner());
+                JSONObject header = new JSONObject();
+                header.putAll(getPath().getHeaders());
+                headersCollection.put(request.getContext().getBppId(),header);
+
+                apiCall.setCallBackPayload(formatter.toString(responseCollection));
+                apiCall.setCallBackHeaders(formatter.toString(headersCollection));
+                apiCall.save();
                 return ack(request);
             }else {
                 return nack(request,request.getContext().getBapId());
             }
         }catch (Exception ex){
             if (request == null){
-                throw new RuntimeException();
+                throw new RuntimeException(ex);
             }
             Request response  = new Request();
             Error error = new Error();
@@ -67,97 +108,50 @@ public class LocalDeliveryBppController extends Controller {
             return new BytesView(getPath(),response.toString().getBytes(StandardCharsets.UTF_8));
         }
     }
-    @RequireLogin(false)
-    public View search() {
-        return act();
-    }
-    @RequireLogin(false)
-    public View select(){
-        return act();
-    }
-    @RequireLogin(false)
-    public View cancel(){
-        return act();
-    }
 
     @RequireLogin(false)
-    public View init(){
+    public View on_search(){
         return act();
     }
     @RequireLogin(false)
-    public View confirm(){
+    public View on_select(){
+        return act();
+    }
+    @RequireLogin(false)
+    public View on_init(){
         return act();
     }
 
     @RequireLogin(false)
-    public View status(){
-        return act();
-    }
-    @RequireLogin(false)
-    public View track(){
-        return act();
-    }
-    @RequireLogin(false)
-    public View update(){
-        return act();
-    }
-
-
-    @RequireLogin(false)
-    public View rating(){
+    public View on_confirm(){
         return act();
     }
 
     @RequireLogin(false)
-    public View support(){
+    public View on_status(){
+        return act();
+    }
+    @RequireLogin(false)
+    public View on_track(){
+        return act();
+    }
+    @RequireLogin(false)
+    public View on_cancel(){
+        return act();
+    }
+    @RequireLogin(false)
+    public View on_update(){
+        return act();
+    }
+
+
+    @RequireLogin(false)
+    public View on_rating(){
         return act();
     }
 
     @RequireLogin(false)
-    public View get_cancellation_reasons(){
-        try {
-            Request request = new Request(StringUtil.read(getPath().getInputStream()));
-            if (request.verifySignature("Authorization",getPath().getHeaders())){
-                Options options = new Options();
-                return new BytesView(getPath(),options.toString().getBytes(StandardCharsets.UTF_8),MimeType.APPLICATION_JSON);
-            }else {
-                return nack(request,request.getContext().getBapId());
-            }
-        }catch (Exception ex){
-            throw new RuntimeException(ex);
-        }
-
+    public View on_support(){
+        return act();
     }
-
-    @RequireLogin(false)
-    public View get_return_reasons(){
-        try {
-            Request request = new Request(StringUtil.read(getPath().getInputStream()));
-            if (request.verifySignature("Authorization",getPath().getHeaders())){
-                Options options = new Options();
-                return new BytesView(getPath(),options.toString().getBytes(StandardCharsets.UTF_8),MimeType.APPLICATION_JSON);
-            }else {
-                return nack(request,request.getContext().getBapId());
-            }
-        }catch (Exception ex){
-            throw new RuntimeException(ex);
-        }
-
-    }
-
-    @RequireLogin(false)
-    public View get_rating_categories(){
-        try {
-            Request request = new Request(StringUtil.read(getPath().getInputStream()));
-            if (request.verifySignature("Authorization",getPath().getHeaders())){
-                return new BytesView(getPath(),new JSONArray().toString().getBytes(StandardCharsets.UTF_8),MimeType.APPLICATION_JSON);
-            }else {
-                return nack(request,request.getContext().getBapId());
-            }
-        }catch (Exception ex){
-            throw new RuntimeException(ex);
-        }
-
-    }
-
 }
