+++ src/main/java/in/succinct/beckn/portal/controller/PostsController.java
@@ -5,10 +5,10 @@ import com.venky.swf.controller.ModelController;
 import com.venky.swf.db.Database;
 import com.venky.swf.db.annotations.column.ui.mimes.MimeType;
 import com.venky.swf.path.Path;
+import com.venky.swf.util.PegDownProcessor;
 import com.venky.swf.views.BytesView;
 import com.venky.swf.views.View;
 import in.succinct.beckn.portal.db.model.collab.Post;
-import com.venky.swf.plugins.wiki.util.PegDownProcessor;
 
 import java.nio.charset.StandardCharsets;
