+++ src/main/resources/templates/governed_areas/index.html
@@ -1,0 +1,1 @@
+Hello Harshit
