+++ src/main/resources/templates/markdown/api_documents.md
@@ -1,0 +1,3 @@
+# Apis 
+Api specifications are available in [github](https://github.com/beckn/protocol-specifications/blob/master/core/v0/api/core.yaml) as [open api specification](https://swagger.io/specification/). 
+
