+++ src/main/resources/templates/markdown/index.md
@@ -1,0 +1,23 @@
+# Welcome to Beckn One (BECKN Open NEwork)
+
+Beckn One is an Open Network which has: 
+
+* A fully functional self serviced registry. 
+* A Gateway.
+* A BPP Testing tool in the form of a BAP. 
+* A BAP Testing tool in the form of a BPP. 
+
+
+Beckn application programmers can use this portal to validate their beckn implementations. 
+
+
+## Quick Links
+1. [Getting Started](getting_started.md)
+2. [Beckn One Network](register_beckn_one.md)
+3. [Api Documentation](api_documents.md)
+4. [The original Beckn White Paper](https://becknprotocol.io/imagining-with-beckn/)
+
+To Register yourself as a beckn application developer with your google id click <a href="/oid/login?SELECTED_OPEN_ID=GOOGLE" target="_blank">here
+    <img src="../images/google-icon.svg"/>
+</a>.
+
